Dragon Curve Resize + Julia HashTables
Dragon curve now random color, resizing works fairly well.  Last iteration still sizes oddly and I'm not sure why.

Did some work on Julia HashTables but I don't know if they're working or not, so you should test that as it'll be easier for you to tell with the higher resolution screen.